Display first swing at the UI
This adds a new Fragment to the UI. The Fragment consists of only a
recyclerView that is controlled by WeatherListAdapter. The
WeatherListAdapter is in charge of inflating and binding the views to
the recyclerview. The first element in the adapter is an expanded row
compared to the rest of the rows and displays the data for today. The
data is held in a data class called WeatherForcast. Not sure if this
builds without some files that haven't been committed yet.