add network and room support
I really hate doing big updates like this, but I've been working on this
in small spurts so I wanted to get thoughts down instead of doing the
right thing.

The main takeaway from this update is the the network fragment and the
room database. The room database starts up and creates a table for our
rest queries. When the network fragment starts up it checks the table
for the last query. If the last query hasnt been in awhile (or the table
is empty) it looks to make a request. The first thing it does is make
sure that the device is connected to a network. Assuming that there is a
connection, it will make an entry in the database with the type of
request (current or forecast) as well as the the api string and a flag
saying that the query is in transit. Once that is complete it will
actually try and send the request. If the request is successfull it will
insert into the DB again this time with the result and a flasg saying
success. Ideally this would be an UPDATE but for some reason room
refuses to actually update the third time. (hopefully can figure this
out before due date but moving on for now).

A few things about the database. Room is pretty nice since it provides
a decent ORM that manages the SQL queries for us.  This part is pretty
straight forward. A small gotcha is that for fields that SQLlite cant
deal with (mainly our enums) you need to add type converters to handle
conversion between ints and enum. Not too complicated.

The archtecture libraries provided by google also give us a live view
class that will handle the MVVM pattern for us. NetworkFragment puts an
observer on this which controls when the timer to do another query are started / canceled.